src/requirements.txt
astroid==2.3.3
atomicwrites==1.3.0
attrs==18.2.0
beautifulsoup4==4.7.1
boto3==1.11.12
botocore==1.14.12
certifi==2018.11.29
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
colorama==0.4.1
decorator==4.3.2
docutils==0.15.2
idna==2.8
ipython-genutils==0.2.0
isort==4.3.21
jmespath==0.9.4
jsonschema==3.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==6.0.0
nbformat==4.4.0
plotly==3.6.1
pluggy==0.9.0
psycopg2==2.8.4
py==1.8.0
pylint==2.4.4
pyrsistent==0.14.11
pytest==4.3.0
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2018.9
PyYAML==3.13
requests==2.21.0
retrying==1.3.3
s3transfer==0.3.3
shodan==1.11.1
six==1.12.0
soupsieve==1.8
texttable==1.6.1
traitlets==4.3.2
typed-ast==1.4.1
urllib3==1.24.1
virtualenv==16.7.9
wrapt==1.11.2
XlsxWriter==1.1.5
